+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
 ## **3.0.0** - *2020-xx-xx*
-* `Dashboard` - app toolbar made responsive;
-* readme updated;
 * dynamic content blocks with Quill WYSIWYG editor in `Dashboard`;
+* `Dashboard`:
+    * app toolbar made responsive;
+    * redirect to `error` page when initial auth request failed;
+* readme updated;
 
 ## **2.3.1** - *2020-04-10*
 * readme updated;